move away from interrupts to speed up the code (clock will go up to approximately 1.5Mhz) (#76)

+++ src/jtag.h
@@ -87,7 +87,7 @@ void jtag_set_srst(uint8_t value);
  * @param in Input data
  * @param out Output data
  */
-void jtag_transfer(uint16_t length, const uint8_t *in,
+void jtag_transfer(uint8_t length, const uint8_t *in,
 		   uint8_t *out);
 
 /**
